added new services, checking student and teacher login later

+++ src/main/java/org/ricardo/school_system/services/StudentService.java
@@ -1,10 +1,15 @@
 package org.ricardo.school_system.services;
 
 import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
+import org.ricardo.school_system.assemblers.StudentForm;
 import org.ricardo.school_system.daos.StudentDao;
 import org.ricardo.school_system.entities.Student;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,13 +18,48 @@ public class StudentService {
 	@Autowired
 	private StudentDao studentDao;
 	
-	public Student add(Student student) {		
-		return studentDao.add(student);
+	@Transactional
+	public ResponseEntity<?> add(HttpServletRequest request, StudentForm studentForm) {		
+		
+		HttpSession session = request.getSession(false);		
+		
+		if (session == null) return new ResponseEntity<>("You are not logged in.", HttpStatus.FORBIDDEN);
+		
+		String[] permissions = (String[]) session.getAttribute("user-permissions");
+		
+		System.out.println("\nCHEGUEI AQUI!\n");
+		
+		for(String permission : permissions) {
+			if(permission.equals("ROLE_TEACHER")) {
+				
+				Student student = new Student(studentForm.getName(), studentForm.getAddress(), 
+											  studentForm.getPhonenumber(), studentForm.getEmail(), 
+											  studentForm.getPassword());
+				
+				studentDao.add(student);
+				
+				return new ResponseEntity<>("Student ´" + student.getName() + "' added.", HttpStatus.OK);
+			}					
+		}
+		
+		return new ResponseEntity<>("You dont´t have enough permissions.", HttpStatus.UNAUTHORIZED);		
 	}
 
 	@Transactional
-	public List<Student> getAll() {
-		return studentDao.getAll();
+	public ResponseEntity<?> getAll(HttpServletRequest request) {
+		
+		HttpSession session = request.getSession(false);		
+		
+		if (session == null) return new ResponseEntity<>("You are not logged in.", HttpStatus.FORBIDDEN);
+		
+		String[] permissions = (String[]) session.getAttribute("user-permissions");
+		
+		for(String permission : permissions) {
+			if(permission.equals("STUDENT"))
+				return new ResponseEntity<>(studentDao.getAll(), HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<>("You dont´t have enough permissions.", HttpStatus.UNAUTHORIZED);	
 	}
 
 	@Transactional
@@ -38,8 +78,22 @@ public class StudentService {
 	}
 
 	@Transactional
-	public void delete(int id) {
-		studentDao.delete(id);
+	public ResponseEntity<?> delete(HttpServletRequest request, int id) {
+		//Nao vai ser o estudante a remover-se....quanto muito vai poder ver o seu perfil. Para já fica assim....
+		HttpSession session = request.getSession(false); 
+		
+		Student sessionStudent = (Student) session.getAttribute("user-credentials");
+		
+		if (session != null && sessionStudent != null) {
+			
+			studentDao.delete(id);	
+	
+			if (sessionStudent.getId() == id) session.invalidate();
+			
+			return new ResponseEntity<>("Student with id " + id + " removed.", HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<>("You are not logged in.", HttpStatus.FORBIDDEN);
 	}
 
 	@Transactional
